excption handling

+++ src/main/java/pl/pjatk/MovieService/MovieService.java
@@ -13,19 +13,32 @@ public class MovieService {
         this.movieJpaRepository = movieJpaRepository;
     }
 
-    public List<Movie> getMovieList() {
+    public List<Movie> getMovieList() throws MovieNotFoundException {
+        if (movieJpaRepository.findAll().isEmpty()) {
+            throw new MovieNotFoundException("Movie list unavailable");
+        }
         return movieJpaRepository.getAllMovies();
     }
 
-    public Movie getMovieByID(Integer id) {
+    public Movie getMovieByID(Integer id) throws MovieNotFoundException {
+        Movie movie = movieJpaRepository.findMovieById(id);
+        if (movie == null) {
+            throw new MovieNotFoundException("Movie not found with ID: " + id);
+        }
         return movieJpaRepository.findMovieById(id);
 
     }
-    public void addMovie(Movie movie) {
+    public void addMovie(Movie movie) throws MovieBadRequestException {
+        if (movie == null || movie.getName() == null || movie.getName().isEmpty()) {
+            throw new MovieBadRequestException("Invalid movie data provided");
+        }
         movieJpaRepository.save(movie);
     }
 
-    public void deleteMovie(Integer id) {
+    public void deleteMovie(Integer id) throws MovieNotFoundException {
+        if (!movieJpaRepository.existsById(id)) {
+            throw new MovieNotFoundException("Movie not found with ID: " + id);
+        }
         movieJpaRepository.deleteById(id);
     }
 }
